% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dkfLLM_weekend_effects_vard.R
\name{dkf.llm.we.vard}
\alias{dkf.llm.we.vard}
\title{Runs the Diffuse Kalman Filter (DFK).
locally linear trend - time varying drift - Saturday and Sunday effects}
\usage{
dkf.llm.we.vard(vec, YY, DOW)
}
\arguments{
\item{vec}{A vector with values of hyperparameters}

\item{YY}{A vector with the values of the time series}

\item{DOW}{A vector with the day of the week, same length as YY}
}
\description{
The state space form is
y(x)  =  Z(x)\emph{alpha(x) + G}u(x)
alpha(x+1) = TT\emph{alpha(x) + H}u(x)
the DKF is initialized with A0 and P0 (Q0=0).
}
\details{
sigma.eps^2 is concentrated out
}
\author{
Sonia Mazzi
}
