% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flightpath.R
\name{flightpath}
\alias{flightpath}
\title{Outputs a connected scatterplot}
\usage{
flightpath(df)
}
\arguments{
\item{df}{The data frame with date, trend and first derivative}
}
\description{
Outputs a connected scatterplot
}
\author{
Sonia Mazzi and Harry Churchley
}
