% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoodLLMvard.R
\name{lik.llm.vard}
\alias{lik.llm.vard}
\title{This the function computes the likelihood for the locally linear state space model with a time varying drift (second component of alpha)}
\usage{
lik.llm.vard(x, y)
}
\arguments{
\item{x}{Values of a vector of hyperparameters}

\item{y}{Vector with the values of the time series}
}
\description{
This the function computes the likelihood for the locally linear state space model with a time varying drift (second component of alpha)
}
\author{
Sonia Mazzi
}
