% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoodLLM_weekend_effects_vard.R
\name{lik.llm.we.vard}
\alias{lik.llm.we.vard}
\title{This is the function that computes the likelihood for LLM-SSM with daily effects with a time varying drift (second component of alpha)
Daily effects are weekdays (5), saturdays (1) and Sundays (1)}
\usage{
lik.llm.we.vard(vec, y, DOW)
}
\arguments{
\item{vec}{A vector with values of hyperparameters}

\item{DOW}{A vector with the day of the week, same length as YY}

\item{YY}{A vector with the values of the time series. Make sure to define externally}
}
\description{
For daily data
}
\author{
Sonia Mazzi
}
