% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_extractor_nowcast.R
\name{trend_extractor_nowcast}
\alias{trend_extractor_nowcast}
\title{Function is used for trend extraction of equally spaced time series data}
\usage{
trend_extractor_nowcast(input_df, nowcast = 0)
}
\arguments{
\item{nowcast}{Whether the data has weekend effects (default is FALSE)

this version has an option for nowcasts: if nowcast = k then the last k observations will be dropped from the data,
k forecasts will be produced using net growth and then added to the data and the trend will be finally extracted from this data}

\item{inpupt_df}{dataframe with date in the first column and the variable of interest in the second column}
}
\description{
Function is used for trend extraction of equally spaced time series data
}
\author{
Sonia Mazzi
}
