% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flightpath.R
\name{flightpath}
\alias{flightpath}
\title{Outputs a connected scatterplot}
\usage{
flightpath(df = randomData, label_type = labeltype)
}
\arguments{
\item{df}{The data frame with date, trend and first derivative}

\item{label_type}{The date format of the time series. There are four options: ymd, md, my, y.}
}
\description{
Outputs a connected scatterplot
}
\author{
Sonia Mazzi and Harry Churchley
}
